requirements.txt
# version independant
pytorch == 2.9.1
torchinfo == 1.8.0
